Add base and attribute parsers
