+++ lib/component_embedded_ruby/parser/base.rb
@@ -1,0 +1,39 @@
+module ComponentEmbeddedRuby
+  class Parser
+    class Base
+      def initialize(token_reader)
+        @token_reader = token_reader
+      end
+
+      private
+
+      def current_token
+        @token_reader.current_token
+      end
+
+      def expect(type)
+        token = current_token
+
+        if token.type != type
+          raise UnexpectedTokenError.new(:string, current_token)
+        else
+          @token_reader.next
+        end
+
+        token
+      end
+
+      def expect_any(*types)
+        token = current_token
+
+        if !types.include?(token.type)
+          raise UnexpectedTokenError.new(:string, token)
+        else
+          @token_reader.next
+        end
+
+        token
+      end
+    end
+  end
+end
